Add a GitHub workflow to build through the build service.
The workflow isn't very specialised - it has been copied from the
template repository with just the names changed to allow it to
build properly.

However, as we're focusing on pull requests, the conditions have
been updated so that it triggers on the pull request not just on
the branch - this is important as otherwise it wouldn't trigger
when we create a PR from a fork.